requirements.txt
aiolimiter==1.0.0
aiosqlite==0.18.0
alembic==1.10.3
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
python-dotenv==1.0.0
python-telegram-bot==20.2
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
